pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zabbix-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Zabbix monitoring platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ori_", email = "oribot@edinprint3d.co.uk"}
]

dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["src.zabbix_mcp"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

[tool.ruff]
line-length = 100
target-version = "py310"
